requirements.txt
# Core dependencies for Unitree Go2 communication
numpy==1.24.3
opencv-python==4.8.0.76
Pillow==10.0.0

# Network communication
websockets==11.0.3
aiohttp==3.9.5
asyncio-mqtt==0.13.0

# Web framework and API
fastapi==0.103.1
uvicorn==0.23.2
jinja2==3.1.2
python-multipart==0.0.6

# Configuration management
pyyaml==6.0.1
python-dotenv==1.0.0

# Data processing and analysis
pandas==2.0.3
scipy==1.11.2

# Logging and monitoring
loguru==0.7.0
psutil==5.9.5

# Testing
pytest==7.4.0
pytest-asyncio==0.21.1
pytest-cov==4.1.0

# Development tools
black==23.7.0
flake8==6.0.0
mypy==1.5.1

# Protocol buffers (for Unitree communication)
protobuf==4.24.0

# Additional utilities
click==8.1.7
rich==13.5.2
tqdm==4.66.1
